requirements.txt
django-modeltranslation